Display itineraries and itinerary summaries
This commit adds displaying of routes and
route summaries on the map page.

Route functionality was refactored
to separate itself from the map related
javascript files since the route/itinerary
stuff was mostly separated from the map.

Styling is still required for the directions;
however, mostly everything is separated into
classes and such so that shouldn't be too
complicated.