Wire up feature directions clicks on explore page
This makes it so the featured destinations that show up in the
left-panel of the explore page can be clicked, which changes
to the directions tab and routes to that destination.

Note: I wasn't able to write a test for this, because the test
suite is currently broken per issue #77.