Clean up map view loading sequence
The initialization of the map view was set up so that both parts (explore
and directions) where initialized, with the active tab defaulting to
'explore' but the 'directions' tab checking the 'method' preference
during its own initialization and grabbing control if it's set to
'directions'.
This moves the choice of initially active tab into the tab controller
initialization and makes the sidebar controllers only run their full
action at initialization if they're active.