Make origin optional in directions view
Peferences.getPreference() normally sets the requested preference to a
default value if it's not set.  This adds an argument to optionally
disable that behavior and uses it to prevent the sidebar directions view
from setting a default origin, to make it possible to link to directions
with a blank origin.