Don't write all parameters to the URL
Reduces the number of preferences that are written to the URL, both
to shorten the URL and because it might not be desirable to force all
selected options in blog post links.

It still reads all possibly-relevant attributes from the URL if they're
there, so we can hand-roll more specific URLs if we want (or, of course,
add front-end functionality in the future).