Upgrade OTP to 0.20.0
Most of the changes here are cleanup in response to Ansible warnings and
such, but the OTP role dependency is updated and the service file
template is changed to align with the setup of the new role version.

Version 1.0.0 of the azavea.opentripplanner role uses Oracle Java 8 and
OTP version 0.20.0.