Use circles for waypoint markers
Leaflet has `Marker`, on which you can very conveniently set
`draggable: true`, and it has `CircleMarker`, which very conveniently
provides circular markers, but ne'er the twain shall meet.

This converts the waypoint markers to CircleMarker then reimplements
dragging using mousedown/mousemove/mouseup events.

The styles for new and existing markers are the same, but I didn't use
a shared options object because I figured there's a good chance we'll
want them to not look identical.

Also the green I chose is nice and Christmasy and won't be confused with
the turn highlighting circles, but possibly a different color scheme is
preferable.

Resolves #522.