Don't close marker under cursor
The add-waypoint marker shows up when you mouse over the line and
disappears after 3 seconds if you don't do anything with it. Except
actually it would disappear 3 seconds after the first mouseout event,
even if that was a momentary flicker or you moused out but came back to
the marker before it disappeared. This makes it save the timeout and
cancel it if the user comes back to the marker.