Add limit param to article endpoint
Resolves #546 by adding support for a `limit` param to the articles
endpoint. So we can get the latest by setting limit=1. If the normal
mode of the endpoint is switched to shuffle, we would probably also want
to add an ordering parameter to preserve the 'latest' behavior, but right
now the default ordering is descending by date published, so the first
is the latest.