Move body div and header/footer into base
Moves the new header and footer into the corresponding mixins and moves
the top-level div into base.html, so that most templates can override
'content' but they don't have to recreate the div or the includes.
Which tab a view belongs under will be supplied by the view function
(at least initially).