Disable URL updating and user preference loading
Disable URL updating as it changes URL to the map page, which we will no longer be using.
Disable loading from user preferences as this will cause the app to jump straight to the
map page when origin and destination are set, and currently there is no way back to the home page.