Closes #547, Closes #548: Implement article card buttons
After some discussion with @lederer, we decided to make the
article preview card be a div with two `<a>` tags to remain HTML5
spec compliant -- interactive elements like links/buttons cannot
be nested within links.

So now the optional detail link is absolutely positioned in
the bottom right corner of the card, and the text margins
were adjusted to give it enough space.

Hooray no javascript!