Closes #581: Fixes place card links
Only the buttons in the place card should be clickable,
unlike the article preview card. The rationale for this,
according to @lederer is:

It’s not obvious (to users, nor to us strategically) which of
those actions the default should be. So we don’t
have a default, and require the user to make an explicit choice.