Fix itinerary plotting
Fix issue with itinerary iteration that was preventing origin/destination marker display.

Lodash forIn iterates over inherited properties as well as elements of the collection.
Replace with forEach to iterate only over the elements, since we only want the itineraries.
Was causing error by attempting to plot a function on the collection object as an itinerary.