Wire up redesigned 'About' page
Converts the About page placeholder to a redesigned view that still uses
the About article from the database.

Note that the existing production About article uses `h3` and `p` tags
whereas the redesign prototype uses `dt` and `dd`. This applies the
styles to both so that either way will work.

The FAQ is gone, but I didn't go so far as to rename the view and model
from 'AboutFaq' to just 'About'. Seems harmless to me and would make it
easier if (heaven forfend) we want a FAQ again.

Resolves  #555.