Incorporate map/directions behavior into home view
Makes the home view check for destination and origin parameters and load
right to the map tab if destination is present or the explore tab if
there's an origin but no destination (note: making the Explore version
of the view work is still on the to-do list). This enables the front-end
URL routing/parameter reading.

Also adds a redirect from `/map/directions` to the home view so that
pre-redesign links to interactive directions will still work.