Convert UserPreferences to in-memory storage
A rough first pass at issue #589. Replaces the storage object in
UserPreferences with one that doesn't use local storage and just keeps
its values in memory. Has the advantage that it doesn't require a huge
refactor of UserPreferences and also would enable a strategy of keeping
some parameters/preferences in local storage and keeping others
ephemeral.

Has the disadvantage that it's kind of weird. I.e. if we don't want local
storage and just want to use URL parameters and function arguments to
control behavior, setting up a UserPreferences controller to mediate
those interactions would not be the most natural approach.