Fix growing-url bug in URL router
Navigo was getting confused about what the root was and tacking URL
updates onto the end rather than replacing. It occurs to me that the URL
setup has gotten simpler such that it might make more sense to skip the
package and do this by hand. But this works for the moment.