Make modal instances use private context
The way the modal controller was set up the global 'options' variable
got updated with each initialization, so all modals because the last
modal loaded.  This makes it private.  I feel like there's probably a
cleaner way to do this than private methods and `bind`, but I didn't
find it and this is working.