Apply flex styles to switchable header content
To make the "Get a Link" modal open when you click that line on the Share
modal, I put the different versions in containing elements that can be
shown and hidden. That made the flex styling break, since flex only
applies to immediate children, so this reproduces the flex styling on
a contents class that can be used inside switchable divs. Still a little
wordy, but better than making multiple `header` elements.