Move origin/dest form into its own controller
Since Directions and Explore both use the origin/destination form, it
needed to handle its own behavior and interact with those controllers
in a more structured way. This makes that happen, moving the form and
typeahead behavior and UserPreferences interactions into a new controller
which interacts with Directions and Explore mainly through events.

Also includes more of the machinery to make the Explore tab and the
interactions between Home, Directions, and Explore work.

Still to do (some carried forward from previous commit):
- The two controllers are still trying to set errors in the form
  themselves, so that needs to be sorted out into form behavior and,
  if any, behavior that Directions or Explore need to take in response
  to changes in error state.
- Activate the link to Explore at the bottom of the Directions sidebar,
  and vice versa
- The places list doesn't change in response to changes in the origin or
  settings. Probably that's the same feature as issue #557. Or it's a
  feature for which that is a special case, since the sidebar list should
  show times corresponding to the selected settings. And probably filter
  out places that are out of range.
- Limit settings available in the trip options modal to ones that are
  relevant
- Trigger update on changes to settings