Move class logic from view to template
If the right body class isn't set on a cold load, it causes a nasty
flash of the wrong styling before the home page controller figures out
what tab to show and the tab control sets the styles. But putting the
whole class string in the django view is gross. This is arguably still
gross, but it makes the view more general and puts the logic for setting
the initial classes slightly closer to the presentation layer.