Remove modeOptionsControl from Explore control
As was done with the Directions controller (56350f91d), instead of
listening to events from the mode options control inside Explore,
expose a `setOptions` method and have the Home view controller listen
for setting schanges and call it.

Also cleared out some commented-out code that seems unlikely to be
needed, and reorganized the function that triggers the isochrone
refresh to move the part that starts the spinner and clears the old
isochrone outside the debounced function, since it doesn't look good
for those things to lag the click.

Also protected Directions.planTrip to only run when the tab is active,
since otherwise the `setOptions` call from the Home controller causes
Directions to seize control.