Set error for place directions with no origin
If someone clicks the directions link on a place card from the Home view
without having entered an origin, it goes to a blank Directions page,
which is probably correct but looks odd. Setting the "from" field to
error and focusing it might make it more obvious what's happening and
what needs to be done.