Highlight marker when mouse is over place card
Restores pre-redesign functionality of highlighting the marker on the map
corresponding to a featured location when you hover over the card for
that location in the sidebar, and highlighting plus centering the map on
a location if you click on its card.

Styling needed for what color the markers should change to and for what
the cards themselves should do when hovered.

With that styling, resolves #728.