Update places on origin/destination reversed
The Explore tab wasn't listening for the `reverse` event from the
directions form, which wasn't having any effect on the isochrone because
it loads origin from preferences when you switch to it, but the nearby
places were going stale. This makes Explore listen for the `reverse`
event and treat it like a typeahead-selected, which causes the nearby
places to be updated.