Only set options from URL if they're non-default
`UserPreferences.isDefault` makes a distinction between options that have
been set by the user to the default value and options that have never
been set. It returns true only for the latter. But since all options get
written to the URL, following a link would result in all options always
seemingly to have been set by the user.

This adds an argument to `UserPreferences.setPreference` to tell it to
ignore the preference if it matches the default value, and updates the
"save options from URL" method to use it.

This will mean that a link copied from a page that was showing the
options banner because a setting had been changed then changed back will
not show the options banner, but that seems better than having links to
pages where no options had been changed always show the options banner.