Make DirectionsList.hide adjust for sidebar banner
The DirectionsList replaces the sidebar banner, but was just restoring
the body class on `hide`, which was fine when there was always a sidebar
banner but doesn't work when there sometimes isn't one. This adds a
selector by which it can figure out if any sidebar banner is active and
set the body-has-sidebar-banner class accordingly.