Enable direct linking to Explore tab
Adds a route to both the back-end and the front-end's URL router to
enable directly linking to the Explore tab. Since the Learn tab is an
independent view, it needs the tab headers to be actual href links.
This enables a '/explore' URL which
- on the back-end, sets the tab context to 'map-explore'
- on the front-end, gets intercepted by the URL router which sets the
'mode' preference to 'explore' then clears the URL (to avoid confusing
later URL manipulation)

Resolves #758.