Use setValue in geolocate
The typeaheads have a function that gets called on `blur` and resets the
value if there's a mismatch between what the field says and the last
selected value. The geolocation callback was setting the field's value
but not using the special `setValue` method that keeps the internal
lastSelectedValue in sync.

This produced no symptoms until PR #787, which caused a blur event to
happen when it wasn't before, which resulted in the value set by the
geolocation being compared to the blank original value, triggering the
reset.

This makes the geolocation callback use `setValue`.