Show geolocation control in map view
For some reason the geolocation control in the origin field
was being hidden in map view. This commit fixes that.

However, said button collides with the reverse-route button. So I
moved the reverse-route button down into the destination field.
It's not optimal but it's a tidy compromise that works.

Resolved #835