Hide events in mobile All filter. Show up to one in map sidebar.
In the desktop home view, we show up to two events at the top
of the two-column place list in the All filter. In mobile home view
and in desktop map view, the place list is a single column, which
makes those top slots extra valuable.

This commit uses the `data-filter` attribute added in #974 to hide
events in the mobile All filter, and to show max one event in the
map sidebar.