Use a shared object for default carousel options
We initialize image carousels in a bunch of places, and they require a big
options object that's mostly the same across all of them. So this makes
a `defaultCarouselOptions` object in `utils.js` and uses that to provide
the repeated options.  The selector is provided in each location (even
though it's usually the same, it seems unwise to centralize that) and the
itinerary slider has some overrides.