Don't repeat the first destination image in the Tour carousel
Showing the first image then loading all the images into the carousel
meant the first one was showing up twice in a row. This adds a slice so
that the first image still shows up before the carousel initializes, but
the carousel images don't include a second copy of it.