Upgrade Troposphere to version 3
This broke majorkirby because it uses `add_version` in its
`set_up_stack()` method, which was changed to set_version() in
Troposphere v3. To work around the issue without pushing a new release
to majorkirby, this removes calls to super() so that the code using
add_version() doesn't get called and then calls set_version() in local
project code so that functionality is preserved.